+++ .circleci/tedana_outputs.txt
@@ -143,7 +143,6 @@ lowk_ts_OC.nii
 meica_mix.1D
 mepca_OC_components.nii
 mepca_mix.1D
-pcastate.pkl
 s0v.nii
 t2sv.nii
 ts_OC.nii
