+++ .circleci/tedana_outputs_verbose.txt
@@ -34,7 +34,6 @@ mepca_S0_pred.nii
 mepca_betas_catd.nii
 mepca_metric_weights.nii
 mepca_mix.1D
-pcastate.pkl
 s0v.nii
 s0vG.nii
 s0vs.nii
