+++ tedana/decomposition/__init__.py
@@ -1,10 +1,8 @@
 # emacs: -*- mode: python-mode; py-indent-offset: 4; tab-width: 4; indent-tabs-mode: nil -*-
 # ex: set sts=4 ts=4 sw=4 et:
 
-from .eigendecomp import (
-    tedpca, tedica,
-)
+from .pca import tedpca
+from .ica import tedica
 
 
-__all__ = [
-    'tedpca', 'tedica']
+__all__ = ['tedpca', 'tedica']
