+++ tedana/selection/__init__.py
@@ -1,8 +1,9 @@
 # emacs: -*- mode: python-mode; py-indent-offset: 4; tab-width: 4; indent-tabs-mode: nil -*-
 # ex: set sts=4 ts=4 sw=4 et:
 
+from .tedpca import kundu_tedpca
 from .select_comps import (
     selcomps,
 )
 
-__all__ = ['selcomps']
+__all__ = ['kundu_tedpca', 'selcomps']
