+++ tedana/workflows/tedana.py
@@ -350,13 +350,13 @@ def tedana_workflow(data, tes, mask=None, mixm=None, ctab=None, manacc=None,
 
     if mixm is None:
         # Identify and remove thermal noise from data
-        n_components, dd = decomposition.tedpca(catd, data_oc, combmode, mask,
+        dd, n_components = decomposition.tedpca(catd, data_oc, combmode, mask,
                                                 t2s, t2sG, ref_img,
                                                 tes=tes, method=tedpca,
                                                 source_tes=source_tes,
                                                 kdaw=10., rdaw=1.,
-                                                verbose=verbose)
-        mmix_orig = decomposition.tedica(n_components, dd, fixed_seed,
+                                                out_dir=out_dir, verbose=verbose)
+        mmix_orig = decomposition.tedica(dd, n_components, fixed_seed,
                                          maxit, maxrestart)
 
         if verbose:
